# contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Play Later API",
    "description": "A gaming backlog management application",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/collection": {
      "get": {
        "tags": [
          "collection"
        ],
        "summary": "List Collection",
        "description": "Get user's game collection with filtering, sorting, and pagination.",
        "operationId": "list_collection_api_v1_collection_get",
        "parameters": [
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by platform",
              "default": null,
              "title": "Platform"
            },
            "description": "Filter by platform"
          },
          {
            "name": "acquisition_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AcquisitionType"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by acquisition type",
              "default": null,
              "title": "Acquisition Type"
            },
            "description": "Filter by acquisition type"
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 5,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by priority (1-5)",
              "default": null,
              "title": "Priority"
            },
            "description": "Filter by priority (1-5)"
          },
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by active status",
              "default": null,
              "title": "Is Active"
            },
            "description": "Filter by active status"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search in game title or notes",
              "default": null,
              "title": "Search"
            },
            "description": "Search in game title or notes"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/CollectionSortBy",
              "description": "Sort field",
              "default": "updated_at"
            },
            "description": "Sort field"
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "description": "Sort order",
              "default": "desc",
              "title": "Sort Order"
            },
            "description": "Sort order"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of items per page",
              "default": 20,
              "title": "Limit"
            },
            "description": "Number of items per page"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip"
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "X-User-Id"
            }
          },
          {
            "name": "session_user",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "Session User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "collection"
        ],
        "summary": "Create Collection Item",
        "description": "Create a new collection item for the authenticated user.",
        "operationId": "create_collection_item_api_v1_collection_post",
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "X-User-Id"
            }
          },
          {
            "name": "session_user",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "Session User"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionItemCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionItemExpanded"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collection/stats": {
      "get": {
        "tags": [
          "collection"
        ],
        "summary": "Get Collection Stats",
        "description": "Get user's collection statistics and insights.",
        "operationId": "get_collection_stats_api_v1_collection_stats_get",
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "X-User-Id"
            }
          },
          {
            "name": "session_user",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "Session User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionStats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collection/{collection_id}": {
      "get": {
        "tags": [
          "collection"
        ],
        "summary": "Get Collection Item",
        "description": "Get a collection item by ID for the authenticated user.",
        "operationId": "get_collection_item_api_v1_collection__collection_id__get",
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Id"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "X-User-Id"
            }
          },
          {
            "name": "session_user",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "Session User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionItemExpanded"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "collection"
        ],
        "summary": "Update Collection Item",
        "description": "Update mutable fields of a collection item for the authenticated user.",
        "operationId": "update_collection_item_api_v1_collection__collection_id__put",
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Id"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "X-User-Id"
            }
          },
          {
            "name": "session_user",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "Session User"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionItemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionItemExpanded"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "collection"
        ],
        "summary": "Delete Collection Item",
        "description": "Delete a collection item (soft delete by default, hard delete with safeguards).",
        "operationId": "delete_collection_item_api_v1_collection__collection_id__delete",
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Id"
            }
          },
          {
            "name": "hard_delete",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Permanently delete the item",
              "default": false,
              "title": "Hard Delete"
            },
            "description": "Permanently delete the item"
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "X-User-Id"
            }
          },
          {
            "name": "session_user",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "Session User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Collection Item Api V1 Collection  Collection Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collection/bulk": {
      "post": {
        "tags": [
          "collection"
        ],
        "summary": "Bulk Collection Operations",
        "description": "Perform bulk operations on multiple collection items.",
        "operationId": "bulk_collection_operations_api_v1_collection_bulk_post",
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "X-User-Id"
            }
          },
          {
            "name": "session_user",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "Session User"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkCollectionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkCollectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcquisitionType": {
        "type": "string",
        "enum": [
          "PHYSICAL",
          "DIGITAL",
          "SUBSCRIPTION",
          "BORROWED",
          "RENTAL"
        ],
        "title": "AcquisitionType"
      },
...
      "BulkCollectionRequest": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/BulkCollectionAction",
            "description": "The action to perform"
          },
          "collection_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Collection Ids",
            "description": "Collection item IDs to update"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Action-specific data"
          }
        },
        "type": "object",
        "required": [
          "action",
          "collection_ids"
        ],
        "title": "BulkCollectionRequest"
      },
...
      "CollectionItemCreate": {
        "properties": {
          "game_id": {
            "type": "string",
            "title": "Game Id",
            "description": "Game ID to add to collection"
          },
          "platform": {
            "type": "string",
            "title": "Platform",
            "description": "Gaming platform"
          },
          "acquisition_type": {
            "$ref": "#/components/schemas/AcquisitionType",
            "description": "How the game was acquired"
          },
          "acquired_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acquired At",
            "description": "When the game was acquired"
          },
          "priority": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 5.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority",
            "description": "Priority level (1-5)"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Personal notes about the game"
          }
        },
        "type": "object",
        "required": [
          "game_id",
          "platform",
          "acquisition_type"
        ],
        "title": "CollectionItemCreate"
      },
      "CollectionItemExpanded": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "game": {
            "$ref": "#/components/schemas/GameDetail"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "acquisition_type": {
            "$ref": "#/components/schemas/AcquisitionType"
          },
          "acquired_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acquired At"
          },
          "priority": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 5.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "playthroughs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Playthroughs"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "game",
          "platform",
          "acquisition_type",
          "created_at",
          "updated_at"
        ],
        "title": "CollectionItemExpanded"
      },
      "CollectionItemUpdate": {
        "properties": {
          "acquisition_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AcquisitionType"
              },
              {
                "type": "null"
              }
            ],
            "description": "How the game was acquired"
          },
          "acquired_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acquired At",
            "description": "When the game was acquired"
          },
          "priority": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 5.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority",
            "description": "Priority level (1-5)"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active",
            "description": "Whether the item is active"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Personal notes about the game"
          }
        },
        "type": "object",
        "title": "CollectionItemUpdate"
      },
      "CollectionListResponse": {
        "properties": {
          "items": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CollectionItemExpanded"
                },
                {
                  "additionalProperties": true,
                  "type": "object"
                }
              ]
            },
            "type": "array",
            "title": "Items"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "filters_applied": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filters Applied"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total_count",
          "limit",
          "offset"
        ],
        "title": "CollectionListResponse"
      },
...
      "CollectionSortBy": {
        "type": "string",
        "enum": [
          "title",
          "acquired_at",
          "priority",
          "platform",
          "updated_at"
        ],
        "title": "CollectionSortBy"
      },
      "CollectionStats": {
        "properties": {
          "total_games": {
            "type": "integer",
            "title": "Total Games"
          },
          "by_platform": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Platform"
          },
          "by_acquisition_type": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Acquisition Type"
          },
          "by_priority": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "By Priority"
          },
          "value_estimate": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "integer"
                    },
                    {
                      "type": "number"
                    },
                    {
                      "type": "string"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Estimate"
          },
          "recent_additions": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recent Additions"
          }
        },
        "type": "object",
        "required": [
          "total_games",
          "by_platform",
          "by_acquisition_type",
          "by_priority"
        ],
        "title": "CollectionStats"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}